requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
jinja2==3.1.4
pillow==10.4.0
